Refactoring, created planned route public DTO
