+++ src/main/java/com/bussin/SpringBack/controllers/DriverController.java
@@ -1,8 +1,9 @@
 package com.bussin.SpringBack.controllers;
 
-import com.bussin.SpringBack.models.Driver;
-import com.bussin.SpringBack.models.DriverDTO;
-import com.bussin.SpringBack.models.PlannedRoute;
+import com.bussin.SpringBack.models.driver.Driver;
+import com.bussin.SpringBack.models.driver.DriverDTO;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoutePublicDTO;
 import com.bussin.SpringBack.services.DriverService;
 import io.swagger.v3.oas.annotations.Operation;
 import lombok.extern.slf4j.Slf4j;
@@ -57,7 +58,7 @@ public class DriverController {
      */
     @Operation
     @GetMapping("/{carPlate}/plannedRoutes")
-    public Set<PlannedRoute> getAllPlannedRoutesByDriver(@Valid @PathVariable String carPlate){
+    public List<PlannedRoutePublicDTO> getAllPlannedRoutesByDriver(@Valid @PathVariable String carPlate){
         log.info(String.format("Retrieving planned routes from driver %s",
                 carPlate));
