+++ src/main/java/com/bussin/SpringBack/models/driver/Driver.java
@@ -1,5 +1,7 @@
-package com.bussin.SpringBack.models;
+package com.bussin.SpringBack.models.driver;
 
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.user.User;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
 import com.fasterxml.jackson.core.JsonProcessingException;
