+++ src/main/java/com/bussin/SpringBack/models/plannedRoute/PlannedRoute.java
@@ -1,5 +1,7 @@
-package com.bussin.SpringBack.models;
+package com.bussin.SpringBack.models.plannedRoute;
 
+import com.bussin.SpringBack.models.ride.Ride;
+import com.bussin.SpringBack.models.driver.Driver;
 import com.fasterxml.jackson.annotation.JsonIdentityInfo;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.ObjectIdGenerators;
@@ -12,6 +14,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
+import lombok.ToString;
 import org.hibernate.annotations.Type;
 
 import javax.persistence.CascadeType;
