+++ src/main/java/com/bussin/SpringBack/models/plannedRoute/PlannedRoutePublicDTO.java
@@ -1,0 +1,21 @@
+package com.bussin.SpringBack.models.plannedRoute;
+
+import com.fasterxml.jackson.annotation.JsonSetter;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+
+@Getter
+@Setter
+@NoArgsConstructor
+@EqualsAndHashCode
+public class PlannedRoutePublicDTO extends PlannedRouteDTO {
+    private String carPlate;
+
+    @JsonSetter
+    public void setCarPlate(String carPlate) {
+        this.carPlate = carPlate;
+    }
+}
