+++ src/main/java/com/bussin/SpringBack/models/user/SignUpUniqueRequest.java
@@ -1,11 +1,10 @@
-package com.bussin.SpringBack.models;
+package com.bussin.SpringBack.models.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.ToString;
 
 import java.io.Serializable;
 import java.util.Objects;
