+++ src/main/java/com/bussin/SpringBack/models/user/SignUpUniqueResponse.java
@@ -1,4 +1,4 @@
-package com.bussin.SpringBack.models;
+package com.bussin.SpringBack.models.user;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
