+++ src/main/java/com/bussin/SpringBack/security/TokenValidator.java
@@ -6,7 +6,7 @@ import com.auth0.jwt.algorithms.Algorithm;
 import com.auth0.jwt.exceptions.JWTVerificationException;
 import com.auth0.jwt.interfaces.DecodedJWT;
 import com.auth0.jwt.interfaces.RSAKeyProvider;
-import com.bussin.SpringBack.models.User;
+import com.bussin.SpringBack.models.user.User;
 import com.bussin.SpringBack.services.UserService;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
