+++ src/main/java/com/bussin/SpringBack/services/RideService.java
@@ -3,10 +3,10 @@ package com.bussin.SpringBack.services;
 import com.bussin.SpringBack.exception.PlannedRouteNotFoundException;
 import com.bussin.SpringBack.exception.RideException;
 import com.bussin.SpringBack.exception.RideNotFoundException;
-import com.bussin.SpringBack.models.PlannedRoute;
-import com.bussin.SpringBack.models.Ride;
-import com.bussin.SpringBack.models.RideDTO;
-import com.bussin.SpringBack.models.User;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.ride.Ride;
+import com.bussin.SpringBack.models.ride.RideDTO;
+import com.bussin.SpringBack.models.user.User;
 import com.bussin.SpringBack.repositories.PlannedRoutesRepository;
 import com.bussin.SpringBack.repositories.RideRepository;
 import org.modelmapper.ModelMapper;
