+++ src/main/java/com/bussin/SpringBack/services/UserService.java
@@ -4,11 +4,11 @@ import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider;
 import com.amazonaws.services.cognitoidp.model.AttributeType;
 import com.amazonaws.services.cognitoidp.model.SignUpRequest;
 import com.bussin.SpringBack.exception.UserNotFoundException;
-import com.bussin.SpringBack.models.SignUpUniqueRequest;
-import com.bussin.SpringBack.models.SignUpUniqueResponse;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserCreationDTO;
-import com.bussin.SpringBack.models.UserDTO;
+import com.bussin.SpringBack.models.user.SignUpUniqueRequest;
+import com.bussin.SpringBack.models.user.SignUpUniqueResponse;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserCreationDTO;
+import com.bussin.SpringBack.models.user.UserDTO;
 import com.bussin.SpringBack.repositories.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
