+++ src/test/java/com/bussin/SpringBack/plannedRouteTests/PlannedRouteServiceTests.java
@@ -2,7 +2,13 @@ package com.bussin.SpringBack.plannedRouteTests;
 
 import com.bussin.SpringBack.TestObjects;
 import com.bussin.SpringBack.exception.PlannedRouteNotFoundException;
-import com.bussin.SpringBack.models.*;
+import com.bussin.SpringBack.models.driver.Driver;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRouteDTO;
+import com.bussin.SpringBack.models.ride.Ride;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserDTO;
+import com.bussin.SpringBack.models.user.UserPublicDTO;
 import com.bussin.SpringBack.repositories.DriverRepository;
 import com.bussin.SpringBack.repositories.PlannedRoutesRepository;
 import com.bussin.SpringBack.services.PlannedRouteService;
