+++ src/test/java/com/bussin/SpringBack/rideTests/RideServiceTests.java
@@ -3,12 +3,12 @@ package com.bussin.SpringBack.rideTests;
 import com.bussin.SpringBack.TestObjects;
 import com.bussin.SpringBack.exception.RideException;
 import com.bussin.SpringBack.exception.RideNotFoundException;
-import com.bussin.SpringBack.models.Driver;
-import com.bussin.SpringBack.models.PlannedRoute;
-import com.bussin.SpringBack.models.Ride;
-import com.bussin.SpringBack.models.RideDTO;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserDTO;
+import com.bussin.SpringBack.models.driver.Driver;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.ride.Ride;
+import com.bussin.SpringBack.models.ride.RideDTO;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserDTO;
 import com.bussin.SpringBack.repositories.PlannedRoutesRepository;
 import com.bussin.SpringBack.repositories.RideRepository;
 import com.bussin.SpringBack.services.RideService;
@@ -24,7 +24,6 @@ import org.modelmapper.PropertyMap;
 import org.springframework.dao.DataIntegrityViolationException;
 
 import javax.validation.ConstraintViolationException;
-import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.Optional;
