+++ src/test/java/com/bussin/SpringBack/userTests/UserCognitoTests.java
@@ -3,8 +3,8 @@ package com.bussin.SpringBack.userTests;
 import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider;
 import com.amazonaws.services.cognitoidp.model.SignUpRequest;
 import com.bussin.SpringBack.TestObjects;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserCreationDTO;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserCreationDTO;
 import com.bussin.SpringBack.repositories.UserRepository;
 import com.bussin.SpringBack.services.UserService;
 import org.junit.jupiter.api.Test;
