+++ src/test/java/com/bussin/SpringBack/userTests/UserCreateIntegrationTests.java
@@ -2,9 +2,9 @@ package com.bussin.SpringBack.userTests;
 
 import com.bussin.SpringBack.TestObjects;
 import com.bussin.SpringBack.integrationTestAuth.CognitoLogin;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserCreationDTO;
-import com.bussin.SpringBack.models.UserDTO;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserCreationDTO;
+import com.bussin.SpringBack.models.user.UserDTO;
 import com.bussin.SpringBack.services.UserService;
 import com.bussin.SpringBack.testConfig.H2JpaConfig;
 import com.bussin.SpringBack.testConfig.TestContextConfig;
