+++ src/test/java/com/bussin/SpringBack/userTests/UserServiceTests.java
@@ -5,12 +5,12 @@ import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProvider;
 import com.amazonaws.services.cognitoidp.AWSCognitoIdentityProviderClientBuilder;
 import com.bussin.SpringBack.TestObjects;
 import com.bussin.SpringBack.exception.UserNotFoundException;
-import com.bussin.SpringBack.models.PlannedRoute;
-import com.bussin.SpringBack.models.Ride;
-import com.bussin.SpringBack.models.SignUpUniqueRequest;
-import com.bussin.SpringBack.models.SignUpUniqueResponse;
-import com.bussin.SpringBack.models.User;
-import com.bussin.SpringBack.models.UserDTO;
+import com.bussin.SpringBack.models.plannedRoute.PlannedRoute;
+import com.bussin.SpringBack.models.ride.Ride;
+import com.bussin.SpringBack.models.user.SignUpUniqueRequest;
+import com.bussin.SpringBack.models.user.SignUpUniqueResponse;
+import com.bussin.SpringBack.models.user.User;
+import com.bussin.SpringBack.models.user.UserDTO;
 import com.bussin.SpringBack.repositories.UserRepository;
 import com.bussin.SpringBack.services.UserService;
 import org.junit.jupiter.api.BeforeEach;
